pyproject.toml
[tool.poetry]
name = "accio"
version = "0.1.0"
description = "Rewrite a federated query into a list of queries with dependencies."
authors = ["Xiaoying Wang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
py4j = "^0.10.9.7"
contexttimer = "^0.3.3"

duckdb = { path = "../postgresscanner/duckdb/tools/pythonpkg", develop = true }
chdb = {path = "../chdb/dist/chdb-1.3.0-cp311-cp311-linux_x86_64.whl"}
polars = {path = "../polars/target/wheels/polars-1.3.0-cp38-abi3-manylinux_2_35_x86_64.whl"}
trino = "^0.329.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pyspark = "^3.5.1"
pandas = "^2.2.2"
pyarrow = "^16.0.0"
numpy = "^1.26.4"
datafusion = "^39.0.0"
connectorx = "^0.3.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
contexttimer = "^0.3.3"
docopt = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
